requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.9.1.post1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
boto3==1.37.37
botocore==1.37.37
CacheControl==0.12.14
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.2
confection==0.1.5
confluent-kafka==2.10.0
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dask==2022.10.2
dateparser==1.2.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fbmessenger==6.0.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.0
grpcio==1.71.0
gTTS==2.5.4
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.0.4
jsonschema==4.17.3
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.12.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mattermostwrapper==2.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==5.2.0
murmurhash==1.0.12
namex==0.0.8
nest-asyncio==1.6.0
networkx==2.6.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.76.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==21.3
pamqp==3.2.1
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pillow==11.2.1
pipenv==2025.0.1
platformdirs==4.3.7
playsound==1.3.0
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.23.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.33.1
pydot==1.4.2
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
PyMuPDF==1.25.5
pyngrok==7.2.5
pyparsing==3.2.3
pyreadline3==3.5.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.11.5
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==4.1.0
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-compat==0.1.3
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.35.0
smart-open==7.1.0
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.13.3
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tokenizers==0.19.1
toolz==1.0.0
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.41.0
twilio==8.2.2
typer==0.15.2
typing-inspection==0.4.0
typing-utils==0.1.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.1
virtualenv==20.30.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xgboost==3.0.2
yarl==1.20.0
zope.interface==7.2
